impl logger scenarios
