+++ test/plugin/scenarios/logger-log4j-scenario/config/expectedData.yaml
@@ -1,0 +1,138 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+segmentItems:
+  - serviceName: logger-log4j-scenario
+    segmentSize: 4
+    segments:
+      - segmentId: 6bda4582407147a78b86e600dfcf7e07.44.16065475238850000
+        spans:
+          - operationName: /log4j/healthCheck
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 1
+            isError: false
+            spanType: Entry
+            peer: ''
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/healthCheck'}
+              - {key: http.method, value: HEAD}
+      - segmentId: not null
+        spans:
+          - operationName: /log4j/no-param
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 1
+            isError: false
+            spanType: Entry
+            peer: ''
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/no-param'}
+              - {key: http.method, value: GET}
+            logs:
+              - logEvent:
+                  - {key: event, value: 'error'}
+                  - {key: message, value: 'no-param'}
+                  - {key: log.kind, value: 'org.apache.skywalking.apm.testcase.logger.controller.CaseController'}
+            refs:
+              - {parentEndpoint: /log4j/testcase, networkAddress: 'localhost:8080', refType: CrossProcess,
+                 parentSpanId: 1, parentTraceSegmentId: not null,
+                 parentServiceInstance: not null, parentService: logger-log4j-scenario,
+                 traceId: not null}
+      - segmentId: not null
+        spans:
+          - operationName: /log4j/one-param
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 1
+            isError: false
+            spanType: Entry
+            peer: ''
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/one-param'}
+              - {key: http.method, value: GET}
+            logs:
+              - logEvent:
+                  - {key: param.1, value: 'java.lang.Exception'}
+                  - {key: event, value: 'error'}
+                  - {key: message, value: 'throwable-param'}
+                  - {key: log.kind, value: 'org.apache.skywalking.apm.testcase.logger.controller.CaseController'}
+            refs:
+              - {parentEndpoint: /log4j/testcase, networkAddress: 'localhost:8080', refType: CrossProcess,
+                 parentSpanId: 2, parentTraceSegmentId: not null,
+                 parentServiceInstance: not null, parentService: logger-log4j-scenario,
+                 traceId: not null}
+      - segmentId: not null
+        spans:
+          - operationName: /log4j/no-param
+            operationId: 0
+            parentSpanId: 0
+            spanId: 1
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 2
+            isError: false
+            spanType: Exit
+            peer: localhost:8080
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/no-param'}
+              - {key: http.method, value: GET}
+          - operationName: /log4j/one-param
+            operationId: 0
+            parentSpanId: 0
+            spanId: 2
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 2
+            isError: false
+            spanType: Exit
+            peer: localhost:8080
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/one-param'}
+              - {key: http.method, value: GET}
+          - operationName: /log4j/testcase
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: nq 0
+            endTime: nq 0
+            componentId: 1
+            isError: false
+            spanType: Entry
+            peer: ''
+            skipAnalysis: false
+            tags:
+              - {key: url, value: 'http://localhost:8080/log4j/testcase'}
+              - {key: http.method, value: GET}
