+++ test/plugin/scenarios/logger-log4j2-scenario/src/main/java/org/apache/skywalking/apm/testcase/logger/Application.java
