+++ test/plugin/scenarios/logger-log4j2-scenario/src/main/java/org/apache/skywalking/apm/testcase/logger/controller/CaseController.java
@@ -1,0 +1,72 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.apm.testcase.logger.controller;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.MarkerManager;
+import org.apache.skywalking.apm.testcase.logger.utils.HttpUtils;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/log4j2")
+public class CaseController {
+
+    private static final Logger logger = LogManager.getLogger(CaseController.class);
+    private static final String SUCCESS = "Success";
+    @Value("${logger.host:localhost:8080}")
+    private String loggerAddress;
+
+    @RequestMapping("/no-param")
+    @ResponseBody
+    public String noParam() {
+        logger.error("no-param");
+        return "no-param";
+    }
+
+    @RequestMapping("/one-param")
+    @ResponseBody
+    public String oneParam() {
+        logger.error("one param is {}", CaseController.class);
+        return "one param";
+    }
+    @RequestMapping("/marker")
+    @ResponseBody
+    public String testMarker() {
+        logger.error(MarkerManager.getMarker("TEST"), "test marker");
+        return "test marker";
+    }
+
+    @RequestMapping("/testcase")
+    public String testcase() {
+        HttpUtils.visit("http://" + loggerAddress + "/log4j2/no-param");
+        HttpUtils.visit("http://" + loggerAddress + "/log4j2/one-param");
+        HttpUtils.visit("http://" + loggerAddress + "/log4j2/marker");
+        return "test";
+    }
+
+    @RequestMapping("/healthCheck")
+    @ResponseBody
+    public String healthCheck() {
+        return SUCCESS;
+    }
+}
