--- test/plugin/scenarios/logger-scenario/logger-projectA-scenario/src/main/java/org/apache/skywalking/apm/testcase/sc/logger/projectA/controller/TestController.java
@@ -1,59 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.apache.skywalking.apm.testcase.sc.logger.projectA.controller;
-
-import org.apache.skywalking.apm.testcase.sc.logger.projectA.utils.HttpUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.io.IOException;
-
-@RestController
-public class TestController {
-
-    @Value("${logback.host:localhost:18080}")
-    private String logbackAddress;
-
-    @Autowired
-    private HttpUtils httpUtils;
-
-    @RequestMapping("/testcase")
-    public String testcase() throws IOException {
-        visit("http://" + logbackAddress + "/logback/no-param");
-        visit("http://" + logbackAddress + "/logback/one-param");
-        visit("http://" + logbackAddress + "/logback/marker");
-        return "test";
-    }
-
-    @RequestMapping("/healthCheck")
-    public String healthCheck() throws IOException {
-        httpUtils.visit("http://" + logbackAddress + "/logback/healthCheck");
-        return "good";
-    }
-
-    private void visit(String path) {
-        try {
-            httpUtils.visit(path);
-        } catch (Exception i) {
-
-        }
-    }
-
-}
